4647 move types to types file and change api (resolve => resolver)
